Update the prepare services and tests
